make set_formnet_resolver function only for linux distros, and set it up to take the tld, so that it can be reused to add other vanity tlds, or potentially real tlds, dns server can handle global lookup through it's fallback mechanism to 8.8.8.8, and we can potentially add a second fallback to 1.1.1.1 in the future
